Fix the bug that rearranged column items
When you declare a col structure in Jeet (i.e. col(1/3)), Jeet will
take the number of columns and use that to generate a nth child
selector to select that last item and float it right.

In our example Jeet would generate .parentcontainer:nth-child(3n) and
float that element to the right.

Jeet will start counting from the very first item in the parent
container. So if you have multiple items in the parent container before
the column items, like a header or paragraph, Jeet will float the items
accordingly.

Lesson learned, wrap your column items in a container div.